feat: adding long usage for SegmentRoundingPolicy

+++ java/memory/memory-core/src/main/java/org/apache/arrow/memory/rounding/SegmentRoundingPolicy.java
@@ -28,7 +28,7 @@ public class SegmentRoundingPolicy implements RoundingPolicy {
    * The segment size. It must be at least {@link SegmentRoundingPolicy#MIN_SEGMENT_SIZE}, and be a
    * power of 2.
    */
-  private int segmentSize;
+  private long segmentSize;
 
   /**
    * Constructor for the segment rounding policy.
@@ -37,7 +37,7 @@ public class SegmentRoundingPolicy implements RoundingPolicy {
    * @throws IllegalArgumentException if the segment size is smaller than {@link
    *     SegmentRoundingPolicy#MIN_SEGMENT_SIZE}, or is not a power of 2.
    */
-  public SegmentRoundingPolicy(int segmentSize) {
+  public SegmentRoundingPolicy(long segmentSize) {
     Preconditions.checkArgument(
         segmentSize >= MIN_SEGMENT_SIZE,
         "The segment size cannot be smaller than %s",
@@ -52,7 +52,7 @@ public class SegmentRoundingPolicy implements RoundingPolicy {
     return (requestSize + (segmentSize - 1)) / segmentSize * segmentSize;
   }
 
-  public int getSegmentSize() {
+  public long getSegmentSize() {
     return segmentSize;
   }
 }
